+++ src/pages/pupil/Dashboard.tsx
@@ -1,4 +1,4 @@
-import { Text, Button, Heading, HStack, useTheme, VStack, useBreakpointValue, Flex, useToast, Alert, Box, Tooltip, Stack } from 'native-base';
+import { Text, Button, HStack, useTheme, VStack, useBreakpointValue, Flex, useToast, Alert, Box, Stack } from 'native-base';
 import { useCallback, useEffect, useMemo, useState } from 'react';
 import AppointmentCard from '../../widgets/AppointmentCard';
 import HSection from '../../widgets/HSection';
@@ -21,7 +21,8 @@ import LearningPartner from '../../widgets/LearningPartner';
 import ImportantInformation from '../../widgets/ImportantInformation';
 import { gql } from '../../gql';
 import HelpNavigation from '../../components/HelpNavigation';
-import { canJoinMeeting } from '../../widgets/appointment/AppointmentDay';
+import NextAppointmentCard from '../../widgets/NextAppointmentCard';
+import { Lecture } from '../../gql/graphql';
 
 type Props = {};
 
@@ -151,7 +152,6 @@ const Dashboard: React.FC<Props> = () => {
     const [showCancelModal, setShowCancelModal] = useState<boolean>(false);
     const [dissolveData, setDissolveData] = useState<{ id: number }>();
     const [toastShown, setToastShown] = useState<boolean>();
-    const [showMeetingNotStarted, setShowMeetingNotStarted] = useState<boolean>();
 
     useEffect(() => {
         trackPageView({
@@ -233,15 +233,6 @@ const Dashboard: React.FC<Props> = () => {
         return data?.me?.pupil?.matches?.filter((match) => !match.dissolved);
     }, [data?.me?.pupil?.matches]);
 
-    const myNextAppointments = useMemo(() => {
-        const nextAppointment = data?.me?.appointments ?? [];
-        const nextAvailableAppointments = nextAppointment.filter((appointment) => {
-            const { start, duration, isOrganizer } = appointment;
-            return canJoinMeeting(start, duration, isOrganizer ? 30 : 10, DateTime.now());
-        });
-        return nextAvailableAppointments.length > 0 ? nextAvailableAppointments : [nextAppointment[0]];
-    }, [data?.me?.appointments]);
-
     return (
         <AsNavigationItem path="start">
             <WithNavigation
@@ -270,55 +261,7 @@ const Dashboard: React.FC<Props> = () => {
                     <VStack paddingX={space['1']} marginX="auto" width="100%" maxWidth={ContainerWidth}>
                         <ImportantInformation variant="dark" />
                         <VStack>
-                            {myNextAppointments && (
-                                <VStack marginBottom={space['1.5']}>
-                                    <Heading marginBottom={space['1']}>{t('dashboard.appointmentcard.header')}</Heading>
-
-                                    <VStack space={space['1']}>
-                                        {myNextAppointments.map((myNextAppointment) => {
-                                            return (
-                                                <AppointmentCard
-                                                    videoButton={
-                                                        <VStack w="100%" space={space['0.5']}>
-                                                            <Tooltip isDisabled={true} maxWidth={300} label={t('course.meeting.hint.pupil')}>
-                                                                <Button
-                                                                    width="100%"
-                                                                    marginTop={space['1']}
-                                                                    onPress={() => {
-                                                                        navigate(`/video-chat/${myNextAppointment.id}/${myNextAppointment.appointmentType}`);
-                                                                    }}
-                                                                    isDisabled={
-                                                                        !myNextAppointment.id ||
-                                                                        !canJoinMeeting(myNextAppointment.start, myNextAppointment.duration, 10, DateTime.now())
-                                                                    }
-                                                                >
-                                                                    {t('course.meeting.videobutton.pupil')}
-                                                                </Button>
-                                                            </Tooltip>
-                                                            {showMeetingNotStarted && <Text color="lightText">{t('course.meeting.videotext')}</Text>}
-                                                        </VStack>
-                                                    }
-                                                    isTeaser={true}
-                                                    onPressToCourse={() => {
-                                                        DateTime.now().plus({ days: 7 }).toISODate();
-                                                        trackEvent({
-                                                            category: 'dashboard',
-                                                            action: 'click-event',
-                                                            name: 'Schüler Dashboard – Termin Teaser | Klick auf' + myNextAppointment.displayName,
-                                                            documentTitle: 'Schüler Dashboard',
-                                                        });
-                                                        navigate(`/appointment/${myNextAppointment.id}`);
-                                                    }}
-                                                    date={myNextAppointment.start}
-                                                    duration={myNextAppointment.duration}
-                                                    title={myNextAppointment.displayName}
-                                                    description={myNextAppointment.description ?? ''}
-                                                />
-                                            );
-                                        })}
-                                    </VStack>
-                                </VStack>
-                            )}
+                            <NextAppointmentCard appointments={data?.me?.appointments as Lecture[]} />
 
                             {/* Matches */}
                             {data?.myRoles?.includes('TUTEE') &&
